Class reorganization on several files
